fixed issue. variable type short does not play well with uint8_t

+++ bitboard_utilities/bitboard-utilities.h
@@ -5,6 +5,6 @@
 
 unsigned lsb_scan(uint64_t bitboard);
 unsigned msb_scan(uint64_t bitboard);
-short EjectIndexFromBitboard(uint64_t &bb, uint64_t *bb_lookup);
+uint8_t EjectIndexFromBitboard(uint64_t &bb, uint64_t *bb_lookup);
 
 #endif
